test: 통합 테스트 전면 개편 및 관련 프로덕션 버그 픽스
- integration 패키지 내 8개 테스트 클래스를 BDD 구조(Given-When-Then)로 전면 개편
- MockMvc를 활용한 API 호출 및 Repository를 통한 DB 직접 검증 로직 도입
- DatabaseCleaner를 적용하여 통합 테스트 간 완벽한 환경 격리 보장
- [Fix] WordDetail 레코드의 필드 타입(difficulty, accuracy)을 long에서 double로 수정하여 QueryDSL 생성자 불일치(500 에러) 해결
- [Fix] WordStatisticJpaEntity 생성 시 @MapsId 충돌을 방지하기 위해 wordId 필드를 null로 설정하여 낙관적 락(ObjectOptimisticLockingFailureException) 문제 해결
- JwtProviderTest 등에서 잘못된 Mocking 구문(anyString) 수정

+++ mobidic-domain/src/main/java/com/kthowns/mobidic/domain/word/model/WordDetail.java
@@ -9,8 +9,8 @@ import java.util.UUID;
 public record WordDetail(
         UUID id,
         String expression,
-        long difficulty,
-        long accuracy,
+        double difficulty,
+        double accuracy,
         boolean isLearned,
         List<Definition> definitions,
         LocalDateTime createdAt
